+++ app/src/main/AndroidManifest.xml
@@ -8,6 +8,7 @@
     <uses-permission android:name="com.google.android.providers.gsf.permission.READ_GSERVICES" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -16,13 +17,6 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyApplication"
        >
-
-        <meta-data
-            android:name="com.google.android.geo.API_KEY"
-            android:value="@string/google_maps_key" />
-        <activity
-            android:name=".MapActivity"
-            android:exported="true" />
         <activity
             android:name=".RegistrationLoginActivity"
             android:exported="true">
