+++ app/src/main/java/com/example/myapplication/screens/login/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.screens
+package com.example.myapplication.screens.login
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,13 +8,14 @@ import android.widget.Button
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.lifecycleScope
 import com.example.myapplication.R
-import com.example.myapplication.data.UserViewModel
+import com.example.myapplication.screens.UserViewModel
 import com.example.myapplication.databinding.FragmentLoginBinding
 import com.google.android.material.textfield.TextInputEditText
 
 
-class LoginFragment:Fragment() {
+class LoginFragment : Fragment() {
 
     lateinit var binding: FragmentLoginBinding
     lateinit var mUserViewModel2: UserViewModel
@@ -48,13 +49,14 @@ class LoginFragment:Fragment() {
             val log = view.findViewById<TextInputEditText>(R.id.log_et_log).text.toString().trim()
             val pass = view.findViewById<TextInputEditText>(R.id.log_et_pass).text.toString().trim()
 
-            mUserViewModel2.getUser(log,pass)
-            val userEntity= mUserViewModel2.user
-            if (userEntity != null) {
-                Toast.makeText(context, "Пользователь найден" , Toast.LENGTH_LONG).show()
-
-            } else {
-                Toast.makeText(context, "Пользователь не найден", Toast.LENGTH_LONG).show()
+            lifecycleScope.launchWhenResumed {
+                mUserViewModel2.getUser(log, pass)
+                val userEntity = mUserViewModel2.user
+                if (userEntity != null) {
+                    Toast.makeText(context, "Пользователь найден", Toast.LENGTH_LONG).show()
+                } else {
+                    Toast.makeText(context, "Пользователь не найден", Toast.LENGTH_LONG).show()
+                }
             }
         }
     }
