+++ app/src/main/res/navigation/nav_graph_application.xml
@@ -9,7 +9,7 @@
     <fragment
 
         android:id="@+id/loginFragment"
-        android:name="com.example.myapplication.screens.LoginFragment"
+        android:name="com.example.myapplication.screens.login.LoginFragment"
         android:label="LoginFragment"
         tools:layout="@layout/fragment_login">
 
@@ -20,7 +20,7 @@
     </fragment>
     <fragment
         android:id="@+id/registrationFragment"
-        android:name="com.example.myapplication.screens.RegistrationFragment"
+        android:name="com.example.myapplication.screens.registration.RegistrationFragment"
         android:label="RegistrationFragment"
         tools:layout="@layout/fragment_registration">
